+++ web-wallet/src/lib/dusk/svelte-stores/index.js
@@ -1,0 +1,1 @@
+export { default as createDataStore } from "./createDataStore";